+++ src/components/navbar/navbar.tsx
@@ -1,10 +1,13 @@
 import { useContext, useEffect, useState } from "react";
-import { Link } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
 import { KeyContext } from "../../App";
 import { MyContextType } from "../../model/MyContextType";
 import logo from "../../assets/logo.png";
 import iconLogin from "../../assets/Vector.png";
 import Carrinho from "../carrinho/Carrinho";
+import Busca from "../../model/Busca";
+import ListarProdutos from "../produtos/ListaProdutos";
+import Produto from "../../model/Produto";
 
 type NavbarProps = {
   onSearch: (searchTerm: string) => void;
@@ -13,13 +16,37 @@ type NavbarProps = {
 const NavBar = () => {
   const { isLoggedIn } = useContext(KeyContext) as MyContextType;
   const [searchTerm, setSearchTerm] = useState("");
+  const [listaProdutos, setListaProdutos] = useState<Produto[]>([]);
+  const [filtro, setFiltro] = useState<string | null>(null);
+  const navigate = useNavigate();
+
+  const buscarProdutos = async (
+    termo: string | null,
+    filtro: string | null,
+    navigate: ReturnType<typeof useNavigate>
+  ) => {
+    const busca: Busca = {
+      pagina: 1,
+      tipoPreco: filtro !== null ? filtro : "MENOR_PRECO",
+      precoMin: null,
+      precoMax: null,
+      categoria: null,
+      vendedor: null,
+      nome: termo !== null && termo.length !== 0 ? termo : null,
+      codigo: null
+    };
+    return await ListarProdutos(busca, navigate);
+  };
 
   const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
     setSearchTerm(event.target.value);
   };
 
   const handleButtonClick = () => {
-    console.log(searchTerm); // Aqui você pode fazer o que quiser com o texto do input
+    (async () =>
+      setListaProdutos(await buscarProdutos(searchTerm, filtro, navigate)))();
+
+      console.log(listaProdutos)
   };
 
   // useEffect(() => {
